moved dist_util to src.utils
