+++ src/eval.py
@@ -6,7 +6,7 @@ from lightning import LightningDataModule, LightningModule, Trainer
 from lightning.pytorch.loggers import Logger
 from omegaconf import DictConfig
 from src.utils import gd_logger
-from src.models.utils import dist_util
+from src.utils import dist_util
 
 rootutils.setup_root(__file__, indicator=".project-root", pythonpath=True)
 # ------------------------------------------------------------------------------------ #
